lesson 10

+++ app/Http/Controllers/PostController.php
@@ -8,12 +8,23 @@ use App\Models\Post;
 use App\Models\Category;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+           new Middleware('auth', except: ['index', 'show'])
+        ];
+    }
+
     public function index(): View
     {
-        $posts = Post::with('category', 'tags')->get();
+        $posts = Post::with('category', 'tags', 'user')->get();
         return view('posts.index', compact('posts'));
     }
 
@@ -38,6 +49,7 @@ class PostController extends Controller
     }
     public function edit(Post $post): View
     {
+        Gate::authorize('postOwner', $post);
         $post->load('tags');
         $categories= Category::all();
         return view('posts.edit',compact('post', 'categories'));
@@ -45,6 +57,7 @@ class PostController extends Controller
     }
     public function update(Post $post, StoreRequest $request, IdArrayRequest $tagsRequest): RedirectResponse
     {
+        Gate::authorize('postOwner', $post);
         $post->update($request->validated());
         $post->tags()->sync($tagsRequest->tags);
         return to_route('posts.index');
@@ -53,7 +66,8 @@ class PostController extends Controller
 
     public function destroy(Post $post): RedirectResponse
     {
-            $post->delete();
+        Gate::authorize('postOwner', $post);
+        $post->delete();
 
         return to_route('posts.index');
     }
